[#11] refactor: extract Codex backend call into src/codex.ts
